Move matlab.frd to frdata module
